+++ Source/CesiumRuntime/Private/Cesium3DTileset.cpp
@@ -15,6 +15,7 @@
 #include "CesiumAsync/CachingAssetAccessor.h"
 #include "CesiumAsync/IAssetResponse.h"
 #include "CesiumAsync/SqliteCache.h"
+#include "CesiumBoundingVolumeComponent.h"
 #include "CesiumCamera.h"
 #include "CesiumCameraManager.h"
 #include "CesiumCommon.h"
@@ -32,6 +33,7 @@
 #include "CesiumRuntimeSettings.h"
 #include "CesiumTextureUtility.h"
 #include "CesiumTransforms.h"
+#include "CesiumViewExtension.h"
 #include "Components/SceneCaptureComponent2D.h"
 #include "CreateGltfOptions.h"
 #include "Engine/Engine.h"
@@ -54,6 +56,8 @@
 #include "Misc/EnumRange.h"
 #include "PhysicsPublicCore.h"
 #include "PixelFormat.h"
+#include "Runtime/Renderer/Private/ScenePrivate.h"
+#include "SceneTypes.h"
 #include "StereoRendering.h"
 #include "UnrealAssetAccessor.h"
 #include "UnrealTaskProcessor.h"
@@ -86,6 +90,8 @@ ACesium3DTileset::ACesium3DTileset()
 
       _lastTilesVisited(0),
       _lastTilesCulled(0),
+      _lastTilesOccluded(0),
+      _lastTilesWaitingForOcclusionResults(0),
       _lastMaxDepthVisited(0),
 
       _captureMovieMode{false},
@@ -215,45 +221,67 @@ void ACesium3DTileset::PostInitProperties() {
 
 void ACesium3DTileset::SetTilesetSource(ETilesetSource InSource) {
   if (InSource != this->TilesetSource) {
-    this->TilesetSource = InSource;
     this->DestroyTileset();
+    this->TilesetSource = InSource;
   }
 }
 
 void ACesium3DTileset::SetUrl(const FString& InUrl) {
   if (InUrl != this->Url) {
-    this->Url = InUrl;
     if (this->TilesetSource == ETilesetSource::FromUrl) {
       this->DestroyTileset();
     }
+    this->Url = InUrl;
   }
 }
 
 void ACesium3DTileset::SetIonAssetID(int64 InAssetID) {
   if (InAssetID >= 0 && InAssetID != this->IonAssetID) {
-    this->IonAssetID = InAssetID;
     if (this->TilesetSource == ETilesetSource::FromCesiumIon) {
       this->DestroyTileset();
     }
+    this->IonAssetID = InAssetID;
   }
 }
 
 void ACesium3DTileset::SetIonAccessToken(const FString& InAccessToken) {
   if (this->IonAccessToken != InAccessToken) {
-    this->IonAccessToken = InAccessToken;
     if (this->TilesetSource == ETilesetSource::FromCesiumIon) {
       this->DestroyTileset();
     }
+    this->IonAccessToken = InAccessToken;
   }
 }
 
 void ACesium3DTileset::SetIonAssetEndpointUrl(
     const FString& InIonAssetEndpointUrl) {
   if (this->IonAssetEndpointUrl != InIonAssetEndpointUrl) {
-    this->IonAssetEndpointUrl = InIonAssetEndpointUrl;
     if (this->TilesetSource == ETilesetSource::FromCesiumIon) {
       this->DestroyTileset();
     }
+    this->IonAssetEndpointUrl = InIonAssetEndpointUrl;
+  }
+}
+
+void ACesium3DTileset::SetEnableOcclusionCulling(bool bEnableOcclusionCulling) {
+  if (this->EnableOcclusionCulling != bEnableOcclusionCulling) {
+    this->EnableOcclusionCulling = bEnableOcclusionCulling;
+    this->DestroyTileset();
+  }
+}
+
+void ACesium3DTileset::SetOcclusionPoolSize(int32 newOcclusionPoolSize) {
+  if (this->OcclusionPoolSize != newOcclusionPoolSize) {
+    this->OcclusionPoolSize = newOcclusionPoolSize;
+    this->DestroyTileset();
+  }
+}
+
+void ACesium3DTileset::SetDelayRefinementForOcclusion(
+    bool bDelayRefinementForOcclusion) {
+  if (this->DelayRefinementForOcclusion != bDelayRefinementForOcclusion) {
+    this->DelayRefinementForOcclusion = bDelayRefinementForOcclusion;
+    this->DestroyTileset();
   }
 }
 
@@ -471,6 +499,11 @@ void ACesium3DTileset::UpdateTransformFromCesium() {
   for (UCesiumGltfComponent* pGltf : gltfComponents) {
     pGltf->UpdateTransformFromCesium(CesiumToUnreal);
   }
+
+  if (this->BoundingVolumePoolComponent) {
+    this->BoundingVolumePoolComponent->UpdateTransformFromCesium(
+        CesiumToUnreal);
+  }
 }
 
 // Called when the game starts or when spawned
@@ -610,7 +643,7 @@ public:
     } else if (pMainThreadResult) {
       UCesiumGltfComponent* pGltf =
           reinterpret_cast<UCesiumGltfComponent*>(pMainThreadResult);
-      this->destroyRecursively(pGltf);
+      CesiumLifetime::destroyComponentRecursively(pGltf);
     }
   }
 
@@ -718,33 +751,6 @@ public:
   }
 
 private:
-  void destroyRecursively(USceneComponent* pComponent) {
-
-    UE_LOG(
-        LogCesium,
-        VeryVerbose,
-        TEXT("Destroying scene component recursively"));
-
-    if (!pComponent) {
-      return;
-    }
-
-    if (pComponent->IsRegistered()) {
-      pComponent->UnregisterComponent();
-    }
-
-    TArray<USceneComponent*> children = pComponent->GetAttachChildren();
-    for (USceneComponent* pChild : children) {
-      this->destroyRecursively(pChild);
-    }
-
-    pComponent->DestroyPhysicsState();
-    pComponent->DestroyComponent();
-    pComponent->ConditionalBeginDestroy();
-
-    UE_LOG(LogCesium, VeryVerbose, TEXT("Destroying scene component done"));
-  }
-
   ACesium3DTileset* _pActor;
 #if PHYSICS_INTERFACE_PHYSX
   IPhysXCookingModule* _pPhysXCookingModule;
@@ -779,6 +785,30 @@ static std::string getCacheDatabaseName() {
   return TCHAR_TO_UTF8(*PlatformAbsolutePath);
 }
 
+void ACesium3DTileset::UpdateLoadStatus() {
+  this->LoadProgress = this->_pTileset->computeLoadProgress();
+
+  if (this->LoadProgress < 100 ||
+      this->_lastTilesWaitingForOcclusionResults > 0) {
+    this->_activeLoading = true;
+  } else if (this->_activeLoading && this->LoadProgress == 100) {
+
+    // There might be a few frames where nothing needs to be loaded as we
+    // are waiting for occlusion results to come back, which means we are not
+    // done with loading all the tiles in the tileset yet.
+    if (this->_lastTilesWaitingForOcclusionResults == 0) {
+
+      // Tileset just finished loading, we broadcast the update
+      UE_LOG(LogCesium, Verbose, TEXT("Broadcasting OnTileLoaded"));
+      OnTilesetLoaded.Broadcast();
+
+      // Tileset remains 100% loaded if we don't have to reload it
+      // so we don't want to keep on sending finished loading updates
+      this->_activeLoading = false;
+    }
+  }
+}
+
 void ACesium3DTileset::LoadTileset() {
   static std::shared_ptr<CesiumAsync::IAssetAccessor> pAssetAccessor =
       std::make_shared<CesiumAsync::CachingAssetAccessor>(
@@ -789,6 +819,9 @@ void ACesium3DTileset::LoadTileset() {
               getCacheDatabaseName()));
   static CesiumAsync::AsyncSystem asyncSystem(
       std::make_shared<UnrealTaskProcessor>());
+  static TSharedRef<CesiumViewExtension, ESPMode::ThreadSafe>
+      cesiumViewExtension =
+          GEngine->ViewExtensions->NewExtension<CesiumViewExtension>();
 
   if (this->_pTileset) {
     // Tileset already loaded, do nothing.
@@ -810,16 +843,41 @@ void ACesium3DTileset::LoadTileset() {
 
   ACesiumCreditSystem* pCreditSystem = this->ResolveCreditSystem();
 
+  this->_cesiumViewExtension = cesiumViewExtension;
+
+  if (this->EnableOcclusionCulling && !this->BoundingVolumePoolComponent) {
+    const glm::dmat4& cesiumToUnreal =
+        GetCesiumTilesetToUnrealRelativeWorldTransform();
+    this->BoundingVolumePoolComponent =
+        NewObject<UCesiumBoundingVolumePoolComponent>(this);
+    this->BoundingVolumePoolComponent->SetUsingAbsoluteLocation(true);
+    this->BoundingVolumePoolComponent->SetFlags(
+        RF_Transient | RF_DuplicateTransient | RF_TextExportTransient);
+    this->BoundingVolumePoolComponent->RegisterComponent();
+    this->BoundingVolumePoolComponent->UpdateTransformFromCesium(
+        cesiumToUnreal);
+  }
+
+  if (this->BoundingVolumePoolComponent) {
+    this->BoundingVolumePoolComponent->initPool(this->OcclusionPoolSize);
+  }
+
   Cesium3DTilesSelection::TilesetExternals externals{
       pAssetAccessor,
       std::make_shared<UnrealResourcePreparer>(this),
       asyncSystem,
       pCreditSystem ? pCreditSystem->GetExternalCreditSystem() : nullptr,
-      spdlog::default_logger()};
+      spdlog::default_logger(),
+      (this->EnableOcclusionCulling && this->BoundingVolumePoolComponent)
+          ? this->BoundingVolumePoolComponent->getPool()
+          : nullptr};
 
   this->_startTime = std::chrono::high_resolution_clock::now();
 
   Cesium3DTilesSelection::TilesetOptions options;
+
+  options.enableOcclusionCulling = this->EnableOcclusionCulling;
+  options.delayRefinementForOcclusion = this->DelayRefinementForOcclusion;
 
   options.showCreditsOnScreen = ShowCreditsOnScreen;
 
@@ -841,7 +899,7 @@ void ACesium3DTileset::LoadTileset() {
             uint8_t(Cesium3DTilesSelection::TilesetLoadType::TilesetJson));
         assert(this->_pTileset == details.pTileset);
 
-        FCesium3DTilesetLoadFailureDetails ueDetails;
+        FCesium3DTilesetLoadFailureDetails ueDetails{};
         ueDetails.Tileset = this;
         ueDetails.Type = ECesium3DTilesetLoadType(typeValue);
         ueDetails.HttpStatusCode =
@@ -931,18 +989,21 @@ void ACesium3DTileset::LoadTileset() {
     }
   }
 
-  if (this->Url.Len() > 0) {
+  switch (this->TilesetSource) {
+  case ETilesetSource::FromUrl:
     UE_LOG(
         LogCesium,
         Log,
         TEXT("Loading tileset from URL %s done"),
         *this->Url);
-  } else {
+    break;
+  case ETilesetSource::FromCesiumIon:
     UE_LOG(
         LogCesium,
         Log,
         TEXT("Loading tileset for asset ID %d done"),
         this->IonAssetID);
+    break;
   }
 
   switch (ApplyDpiScaling) {
@@ -962,19 +1023,25 @@ void ACesium3DTileset::LoadTileset() {
 }
 
 void ACesium3DTileset::DestroyTileset() {
+  if (this->_cesiumViewExtension) {
+    this->_cesiumViewExtension = nullptr;
+  }
 
-  if (this->Url.Len() > 0) {
+  switch (this->TilesetSource) {
+  case ETilesetSource::FromUrl:
     UE_LOG(
         LogCesium,
         Verbose,
         TEXT("Destroying tileset from URL %s"),
         *this->Url);
-  } else {
+    break;
+  case ETilesetSource::FromCesiumIon:
     UE_LOG(
         LogCesium,
         Verbose,
         TEXT("Destroying tileset for asset ID %d"),
         this->IonAssetID);
+    break;
   }
 
   // The way CesiumRasterOverlay::add is currently implemented, destroying the
@@ -995,18 +1062,21 @@ void ACesium3DTileset::DestroyTileset() {
 
   this->_pTileset.Reset();
 
-  if (this->Url.Len() > 0) {
+  switch (this->TilesetSource) {
+  case ETilesetSource::FromUrl:
     UE_LOG(
         LogCesium,
         Verbose,
         TEXT("Destroying tileset from URL %s done"),
         *this->Url);
-  } else {
+    break;
+  case ETilesetSource::FromCesiumIon:
     UE_LOG(
         LogCesium,
         Verbose,
         TEXT("Destroying tileset for asset ID %d done"),
         this->IonAssetID);
+    break;
   }
 }
 
@@ -1501,8 +1571,10 @@ void ACesium3DTileset::updateTilesetOptionsFromProperties() {
   options.forbidHoles = this->ForbidHoles;
   options.maximumSimultaneousTileLoads = this->MaximumSimultaneousTileLoads;
   options.loadingDescendantLimit = this->LoadingDescendantLimit;
-
   options.enableFrustumCulling = this->EnableFrustumCulling;
+  options.enableOcclusionCulling = this->EnableOcclusionCulling;
+
+  options.delayRefinementForOcclusion = this->DelayRefinementForOcclusion;
   options.enableFogCulling = this->EnableFogCulling;
   options.enforceCulledScreenSpaceError = this->EnforceCulledScreenSpaceError;
   options.culledScreenSpaceError =
@@ -1523,6 +1595,9 @@ void ACesium3DTileset::updateLastViewUpdateResultState(
       result.tilesVisited != this->_lastTilesVisited ||
       result.culledTilesVisited != this->_lastCulledTilesVisited ||
       result.tilesCulled != this->_lastTilesCulled ||
+      result.tilesOccluded != this->_lastTilesOccluded ||
+      result.tilesWaitingForOcclusionResults !=
+          this->_lastTilesWaitingForOcclusionResults ||
       result.maxDepthVisited != this->_lastMaxDepthVisited) {
 
     this->_lastTilesRendered = result.tilesToRenderThisFrame.size();
@@ -1533,13 +1608,16 @@ void ACesium3DTileset::updateLastViewUpdateResultState(
     this->_lastTilesVisited = result.tilesVisited;
     this->_lastCulledTilesVisited = result.culledTilesVisited;
     this->_lastTilesCulled = result.tilesCulled;
+    this->_lastTilesOccluded = result.tilesOccluded;
+    this->_lastTilesWaitingForOcclusionResults =
+        result.tilesWaitingForOcclusionResults;
     this->_lastMaxDepthVisited = result.maxDepthVisited;
 
     UE_LOG(
         LogCesium,
         Display,
         TEXT(
-            "%s: %d ms, Visited %d, Culled Visited %d, Rendered %d, Culled %d, Max Depth Visited: %d, Loading-Low %d, Loading-Medium %d, Loading-High %d"),
+            "%s: %d ms, Visited %d, Culled Visited %d, Rendered %d, Culled %d, Occluded %d, Waiting For Occlusion Results %d, Max Depth Visited: %d, Loading-Low %d, Loading-Medium %d, Loading-High %d, Loaded tiles %g%%"),
         *this->GetName(),
         (std::chrono::high_resolution_clock::now() - this->_startTime).count() /
             1000000,
@@ -1547,10 +1625,13 @@ void ACesium3DTileset::updateLastViewUpdateResultState(
         result.culledTilesVisited,
         result.tilesToRenderThisFrame.size(),
         result.tilesCulled,
+        result.tilesOccluded,
+        result.tilesWaitingForOcclusionResults,
         result.maxDepthVisited,
         result.tilesLoadingLowPriority,
         result.tilesLoadingMediumPriority,
-        result.tilesLoadingHighPriority);
+        result.tilesLoadingHighPriority,
+        this->LoadProgress);
   }
 }
 
@@ -1638,6 +1719,21 @@ void ACesium3DTileset::Tick(float DeltaTime) {
     }
   }
 
+  if (this->BoundingVolumePoolComponent && this->_cesiumViewExtension) {
+    const TArray<USceneComponent*>& children =
+        this->BoundingVolumePoolComponent->GetAttachChildren();
+    for (USceneComponent* pChild : children) {
+      UCesiumBoundingVolumeComponent* pBoundingVolume =
+          Cast<UCesiumBoundingVolumeComponent>(pChild);
+
+      if (!pBoundingVolume) {
+        continue;
+      }
+
+      pBoundingVolume->UpdateOcclusion(*this->_cesiumViewExtension.Get());
+    }
+  }
+
   updateTilesetOptionsFromProperties();
 
   std::vector<FCesiumCamera> cameras = this->GetCameras();
@@ -1658,6 +1754,7 @@ void ACesium3DTileset::Tick(float DeltaTime) {
       this->_captureMovieMode ? this->_pTileset->updateViewOffline(frustums)
                               : this->_pTileset->updateView(frustums);
   updateLastViewUpdateResultState(result);
+  this->UpdateLoadStatus();
 
   removeVisibleTilesFromList(
       this->_tilesToNoLongerRenderNextFrame,
@@ -1725,6 +1822,8 @@ void ACesium3DTileset::PostEditChangeProperty(
       PropName == GET_MEMBER_NAME_CHECKED(ACesium3DTileset, Material) ||
       PropName == GET_MEMBER_NAME_CHECKED(ACesium3DTileset, WaterMaterial) ||
       PropName == GET_MEMBER_NAME_CHECKED(ACesium3DTileset, ApplyDpiScaling) ||
+      PropName ==
+          GET_MEMBER_NAME_CHECKED(ACesium3DTileset, EnableOcclusionCulling) ||
       // For properties nested in structs, GET_MEMBER_NAME_CHECKED will prefix
       // with the struct name, so just do a manual string comparison.
       PropNameAsString == TEXT("RenderCustomDepth") ||